// src/lib/grpc/gen/prosper/v1/ledger.ts
export interface BankAccount {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: int32 bank_id = 3
     */
    bankId: number;
    /**
     * Either currency_code or stock_id is set, never both.
     *
     * @generated from protobuf field: optional string currency_code = 4
     */
    currencyCode?: string;
    /**
     * @generated from protobuf field: optional int32 stock_id = 5
     */
    stockId?: number;
    /**
     * @generated from protobuf field: bool joint = 6
     */
    joint: boolean;
    /**
     * @generated from protobuf field: bool archived = 7
     */
    archived: boolean;
    /**
     * @generated from protobuf field: int32 display_order = 8
     */
    displayOrder: number;
    /**
     * @generated from protobuf field: int32 initial_balance_cents = 9
     */
    initialBalanceCents: number;
}
// ...
export interface AccountUnit {
    /**
     * @generated from protobuf oneof: unit
     */
    unit: {
        oneofKind: "currencyCode";
        /**
         * @generated from protobuf field: string currency_code = 1
         */
        currencyCode: string;
    } | {
        oneofKind: "stockId";
        /**
         * @generated from protobuf field: int32 stock_id = 2
         */
        stockId: number;
    } | {
        oneofKind: "newStock";
        /**
         * @generated from protobuf field: prosper.v1.StockSpec new_stock = 3
         */
        newStock: StockSpec;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message prosper.v1.StockSpec
 */
export interface StockSpec {
    /**
     * @generated from protobuf field: string exchange = 1
     */
    exchange: string;
    /**
     * @generated from protobuf field: string ticker = 2
     */
    ticker: string;
}
// ...
export const BankAccount = new BankAccount$Type();
// ...
class AccountUnit$Type extends MessageType<AccountUnit> {
    constructor() {
        super("prosper.v1.AccountUnit", [
            { no: 1, name: "currency_code", kind: "scalar", oneof: "unit", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "stock_id", kind: "scalar", oneof: "unit", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "new_stock", kind: "message", oneof: "unit", T: () => StockSpec }
        ]);
    }
    create(value?: PartialMessage<AccountUnit>): AccountUnit {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.unit = { oneofKind: undefined };
        if (value !== undefined)
            reflectionMergePartial<AccountUnit>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: AccountUnit): AccountUnit {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string currency_code */ 1:
                    message.unit = {
                        oneofKind: "currencyCode",
                        currencyCode: reader.string()
                    };
                    break;
                case /* int32 stock_id */ 2:
                    message.unit = {
                        oneofKind: "stockId",
                        stockId: reader.int32()
                    };
                    break;
                case /* prosper.v1.StockSpec new_stock */ 3:
                    message.unit = {
                        oneofKind: "newStock",
                        newStock: StockSpec.internalBinaryRead(reader, reader.uint32(), options, (message.unit as any).newStock)
                    };
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: AccountUnit, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string currency_code = 1; */
        if (message.unit.oneofKind === "currencyCode")
            writer.tag(1, WireType.LengthDelimited).string(message.unit.currencyCode);
        /* int32 stock_id = 2; */
        if (message.unit.oneofKind === "stockId")
            writer.tag(2, WireType.Varint).int32(message.unit.stockId);
        /* prosper.v1.StockSpec new_stock = 3; */
        if (message.unit.oneofKind === "newStock")
            StockSpec.internalBinaryWrite(message.unit.newStock, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message prosper.v1.AccountUnit
 */
export const AccountUnit = new AccountUnit$Type();
// @generated message type with reflection information, may provide speed optimized methods
class StockSpec$Type extends MessageType<StockSpec> {
    constructor() {
        super("prosper.v1.StockSpec", [
            { no: 1, name: "exchange", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "ticker", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<StockSpec>): StockSpec {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.exchange = "";
        message.ticker = "";
        if (value !== undefined)
            reflectionMergePartial<StockSpec>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: StockSpec): StockSpec {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string exchange */ 1:
                    message.exchange = reader.string();
                    break;
                case /* string ticker */ 2:
                    message.ticker = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: StockSpec, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string exchange = 1; */
        if (message.exchange !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.exchange);
        /* string ticker = 2; */
        if (message.ticker !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.ticker);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message prosper.v1.StockSpec
 */
export const StockSpec = new StockSpec$Type();